Add logic for displaying data from local storage
Helper function "displayContent()" is responsible for handling two possible
scenarios. The first scenario is that there is no data in local storage and in
that case, no Table is shown. The second scenario is that there is data in local
storage. Which data will be displayed in the Table depends on whether the email
from local storage data is the same as email that is currently logged in.